nuc64: auto generate

+++ speedtest1/ccgo_linux_amd64.go
@@ -2531,7 +2531,7 @@ func HashInit(tls *libc.TLS) {
 		if !(k < uint32(256)) {
 			break
 		}
-		*(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(k))) = uint8(uint8(k))
+		*(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(k))) = uint8(k)
 		goto _1
 	_1:
 		k++
@@ -2551,7 +2551,7 @@ func HashUpdate(tls *libc.TLS, aData uintptr, nData uint32) {
 	i = g.hash.i
 	j = g.hash.j
 	if g.hashFile != 0 {
-		libc.Xfwrite(tls, aData, uint64(1), uint64(uint64(nData)), g.hashFile)
+		libc.Xfwrite(tls, aData, uint64(1), uint64(nData), g.hashFile)
 	}
 	k = uint32(0)
 	for {
@@ -2590,7 +2590,7 @@ func HashFinal(tls *libc.TLS) {
 		}
 		i++
 		t = *(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(i)))
-		j = uint8(int32(j) + int32(int32(t)))
+		j = uint8(int32(j) + int32(t))
 		*(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(i))) = *(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(j)))
 		*(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(j))) = t
 		t = uint8(int32(t) + int32(*(*uint8)(unsafe.Pointer(uintptr(unsafe.Pointer(&g)) + 3152 + 3 + uintptr(i)))))
@@ -2611,14 +2611,14 @@ func HashFinal(tls *libc.TLS) {
 //	** is not a hex digit.
 //	*/
 func hexDigitValue(tls *libc.TLS, c int8) (r int32) {
-	if int32(int32(c)) >= int32('0') && int32(int32(c)) <= int32('9') {
-		return int32(int32(c)) - int32('0')
+	if int32(c) >= int32('0') && int32(c) <= int32('9') {
+		return int32(c) - int32('0')
 	}
-	if int32(int32(c)) >= int32('a') && int32(int32(c)) <= int32('f') {
-		return int32(int32(c)) - int32('a') + int32(10)
+	if int32(c) >= int32('a') && int32(c) <= int32('f') {
+		return int32(c) - int32('a') + int32(10)
 	}
-	if int32(int32(c)) >= int32('A') && int32(int32(c)) <= int32('F') {
-		return int32(int32(c)) - int32('A') + int32(10)
+	if int32(c) >= int32('A') && int32(c) <= int32('F') {
+		return int32(c) - int32('A') + int32(10)
 	}
 	return -int32(1)
 }
@@ -2654,7 +2654,7 @@ func integerValue(tls *libc.TLS, zArg uintptr) (r int32) {
 			if !(v1 >= 0) {
 				break
 			}
-			v = v<<libc.Int32FromInt32(4) + int64(int64(x))
+			v = v<<libc.Int32FromInt32(4) + int64(x)
 			zArg++
 		}
 	} else {
@@ -2665,7 +2665,7 @@ func integerValue(tls *libc.TLS, zArg uintptr) (r int32) {
 	}
 	i = 0
 	for {
-		if !(uint64(uint64(i)) < libc.Uint64FromInt64(144)/libc.Uint64FromInt64(16)) {
+		if !(uint64(i) < libc.Uint64FromInt64(144)/libc.Uint64FromInt64(16)) {
 			break
 		}
 		if libsqlite3.Xsqlite3_stricmp(tls, aMult[i].zSuffix, zArg) == 0 {
@@ -3205,7 +3205,7 @@ func speedtest1_run(tls *libc.TLS) {
 						}
 						g.nResByte += uint64(nBlob*int32(2) + int32(2))
 					} else {
-						HashUpdate(tls, z1, uint32(uint32(len1)))
+						HashUpdate(tls, z1, uint32(len1))
 						g.nResByte += uint64(len1 + int32(2))
 					}
 				}
@@ -3302,7 +3302,7 @@ func testset_main(tls *libc.TLS) {
 	n = v1
 	sz = v1
 	(*(*[2000]int8)(unsafe.Pointer(bp)))[0] = 0
-	maxb = int32(roundup_allones(tls, uint32(uint32(sz))))
+	maxb = int32(roundup_allones(tls, uint32(sz)))
 	speedtest1_begin_test(tls, int32(100), __ccgo_ts+590, libc.VaList(bp+2008, n))
 	speedtest1_exec(tls, __ccgo_ts+626, 0)
 	speedtest1_exec(tls, __ccgo_ts+632, libc.VaList(bp+2008, isTemp(tls, int32(9)), g.zNN, g.zNN, g.zNN))
@@ -3312,9 +3312,9 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
 		speedtest1_numbername(tls, x1, bp, int32(2000))
-		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(1), int64(int64(x1)))
+		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(1), int64(x1))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), i)
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(3), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
@@ -3334,10 +3334,10 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
 		speedtest1_numbername(tls, x1, bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), i)
-		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(2), int64(int64(x1)))
+		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(2), int64(x1))
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(3), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
 		goto _3
@@ -3356,10 +3356,10 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
 		speedtest1_numbername(tls, x1, bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), i)
-		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(1), int64(int64(x1)))
+		libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(1), int64(x1))
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(3), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
 		goto _4
@@ -3378,11 +3378,11 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			x2 = speedtest1_random(tls)%uint32(10) + uint32(sz/int32(5000)) + x1
 		}
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _5
 	_5:
@@ -3400,9 +3400,9 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[0] = int8('%')
-			len1 = speedtest1_numbername(tls, uint32(uint32(i)), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
+			len1 = speedtest1_numbername(tls, uint32(i), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1] = int8('%')
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1+int32(1)] = 0
 		}
@@ -3424,9 +3424,9 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[0] = int8('%')
-			len1 = speedtest1_numbername(tls, uint32(uint32(i)), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
+			len1 = speedtest1_numbername(tls, uint32(i), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1] = int8('%')
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1+int32(1)] = 0
 		}
@@ -3448,9 +3448,9 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[0] = int8('%')
-			len1 = speedtest1_numbername(tls, uint32(uint32(i)), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
+			len1 = speedtest1_numbername(tls, uint32(i), bp+uintptr(1), int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(2)))
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1] = int8('%')
 			(*(*[2000]int8)(unsafe.Pointer(bp)))[len1+int32(1)] = 0
 		}
@@ -3481,11 +3481,11 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			x2 = speedtest1_random(tls)%uint32(10) + uint32(sz/int32(5000)) + x1
 		}
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _9
 	_9:
@@ -3503,11 +3503,11 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+			x1 = speedtest1_random(tls) % uint32(maxb)
 			x2 = speedtest1_random(tls)%uint32(10) + uint32(sz/int32(5000)) + x1
 		}
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _10
 	_10:
@@ -3525,7 +3525,7 @@ func testset_main(tls *libc.TLS) {
 			break
 		}
 		if (i-int32(1))%g.nRepeat == 0 {
-			x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+			x1 = swizzle(tls, uint32(i), uint32(maxb))
 			len1 = speedtest1_numbername(tls, x1, bp, int32(libc.Uint64FromInt64(2000)-libc.Uint64FromInt32(1)))
 		}
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(1), bp, len1, libc.UintptrFromInt32(0))
@@ -3566,10 +3566,10 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = speedtest1_random(tls) % uint32(uint32(maxb))
+		x1 = speedtest1_random(tls) % uint32(maxb)
 		x2 = speedtest1_random(tls)%uint32(10) + uint32(sz/int32(5000)) + x1
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _12
 	_12:
@@ -3586,8 +3586,8 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = speedtest1_random(tls)%uint32(uint32(sz)) + uint32(1)
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
+		x1 = speedtest1_random(tls)%uint32(sz) + uint32(1)
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
 		speedtest1_run(tls)
 		goto _13
 	_13:
@@ -3610,10 +3610,10 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = speedtest1_random(tls)%uint32(uint32(maxb)) + uint32(1)
+		x1 = speedtest1_random(tls)%uint32(maxb) + uint32(1)
 		x2 = speedtest1_random(tls)%uint32(10) + uint32(sz/int32(5000)) + x1
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _14
 	_14:
@@ -3630,8 +3630,8 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = speedtest1_random(tls)%uint32(uint32(sz)) + uint32(1)
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
+		x1 = speedtest1_random(tls)%uint32(sz) + uint32(1)
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
 		speedtest1_run(tls)
 		goto _15
 	_15:
@@ -3657,10 +3657,10 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = speedtest1_random(tls)%uint32(uint32(sz)) + uint32(1)
+		x1 = speedtest1_random(tls)%uint32(sz) + uint32(1)
 		x2 = speedtest1_random(tls)%uint32(10) + x1 + uint32(4)
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(int32(x2)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), int32(x2))
 		speedtest1_run(tls)
 		goto _16
 	_16:
@@ -3687,9 +3687,9 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
-		speedtest1_numbername(tls, uint32(uint32(i)), bp, int32(2000))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int64(int64(x1))))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
+		speedtest1_numbername(tls, uint32(i), bp, int32(2000))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int64(x1)))
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(2), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
 		goto _18
@@ -3714,8 +3714,8 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int64(int64(x1))))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int64(x1)))
 		speedtest1_run(tls)
 		goto _19
 	_19:
@@ -3744,7 +3744,7 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
 		speedtest1_numbername(tls, x1, bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(2), i)
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(1), bp, -int32(1), libc.UintptrFromInt32(0))
@@ -3766,7 +3766,7 @@ func testset_main(tls *libc.TLS) {
 		if !(i <= n) {
 			break
 		}
-		x1 = swizzle(tls, uint32(uint32(i)), uint32(uint32(maxb)))
+		x1 = swizzle(tls, uint32(i), uint32(maxb))
 		speedtest1_numbername(tls, x1, bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(1), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
@@ -3983,13 +3983,13 @@ func testset_orm(tls *libc.TLS) {
 			case int32('I'):
 				fallthrough
 			case int32('T'):
-				libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(j+uint32(2)), int64(int64(x1)))
+				libsqlite3.Xsqlite3_bind_int64(tls, g.pStmt, int32(j+uint32(2)), int64(x1))
 			case int32('F'):
-				libsqlite3.Xsqlite3_bind_double(tls, g.pStmt, int32(j+uint32(2)), float64(float64(x1)))
+				libsqlite3.Xsqlite3_bind_double(tls, g.pStmt, int32(j+uint32(2)), float64(x1))
 			case int32('V'):
 				fallthrough
 			case int32('B'):
-				libsqlite3.Xsqlite3_bind_text64(tls, g.pStmt, int32(j+uint32(2)), bp, uint64(uint64(len1)), libc.UintptrFromInt32(0), uint8(SQLITE_UTF8))
+				libsqlite3.Xsqlite3_bind_text64(tls, g.pStmt, int32(j+uint32(2)), bp, uint64(len1), libc.UintptrFromInt32(0), uint8(SQLITE_UTF8))
 				break
 			}
 			goto _3
@@ -4012,7 +4012,7 @@ func testset_orm(tls *libc.TLS) {
 			break
 		}
 		x1 = speedtest1_random(tls) % nRow
-		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(int32(x1)))
+		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), int32(x1))
 		speedtest1_run(tls)
 		goto _4
 	_4:
@@ -4047,8 +4047,8 @@ func testset_trigger(tls *libc.TLS) {
 			if !(ii < NROW) {
 				break
 			}
-			x1 = int32(speedtest1_random(tls) % uint32(uint32(NROW)))
-			speedtest1_numbername(tls, uint32(uint32(x1)), bp, int32(2000))
+			x1 = int32(speedtest1_random(tls) % uint32(NROW))
+			speedtest1_numbername(tls, uint32(x1), bp, int32(2000))
 			libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), x1)
 			libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(2), bp, -int32(1), libc.UintptrFromInt32(0))
 			speedtest1_run(tls)
@@ -4180,7 +4180,7 @@ func testset_trigger(tls *libc.TLS) {
 		if !(jj < NROW2) {
 			break
 		}
-		speedtest1_numbername(tls, uint32(uint32(jj)), bp, int32(2000))
+		speedtest1_numbername(tls, uint32(jj), bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), jj)
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(2), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
@@ -4243,7 +4243,7 @@ func testset_trigger(tls *libc.TLS) {
 		if !(jj < NROW2) {
 			break
 		}
-		speedtest1_numbername(tls, uint32(uint32(jj)), bp, int32(2000))
+		speedtest1_numbername(tls, uint32(jj), bp, int32(2000))
 		libsqlite3.Xsqlite3_bind_int(tls, g.pStmt, int32(1), jj)
 		libsqlite3.Xsqlite3_bind_text(tls, g.pStmt, int32(2), bp, -int32(1), libc.UintptrFromInt32(0))
 		speedtest1_run(tls)
@@ -4375,11 +4375,11 @@ func displayLinuxIoStats(tls *libc.TLS, out uintptr) {
 	for libc.Xfgets(tls, bp, int32(200), in) != uintptr(0) {
 		i = 0
 		for {
-			if !(uint64(uint64(i)) < libc.Uint64FromInt64(112)/libc.Uint64FromInt64(16)) {
+			if !(uint64(i) < libc.Uint64FromInt64(112)/libc.Uint64FromInt64(16)) {
 				break
 			}
 			n = int32(libc.Xstrlen(tls, aTrans[i].zPattern))
-			if libc.Xstrncmp(tls, aTrans[i].zPattern, bp, uint64(uint64(n))) == 0 {
+			if libc.Xstrncmp(tls, aTrans[i].zPattern, bp, uint64(n)) == 0 {
 				libc.Xfprintf(tls, out, __ccgo_ts+19321, libc.VaList(bp+208, aTrans[i].zDesc, bp+uintptr(n)))
 				break
 			}
@@ -4466,9 +4466,9 @@ func main1(tls *libc.TLS, argc int32, argv uintptr) (r int32) {
 	doTrace = 0                                                                                      /* True for --trace */
 	zEncoding = uintptr(0)                                                                           /* --utf16be or --utf16le */
 	zDbName = uintptr(0)                                                                             /* Name of the test database */
-	pHeap = uintptr(0)   /* Allocated heap space */
-	pLook = uintptr(0)   /* Allocated lookaside space */
-	pPCache = uintptr(0) /* API return code */
+	pHeap = uintptr(0)                                                                               /* Allocated heap space */
+	pLook = uintptr(0)                                                                               /* Allocated lookaside space */
+	pPCache = uintptr(0)                                                                             /* API return code */
 	/*
 	 ** Confirms that argc has at least N arguments following argv[i]. */
 	/* Display the version of SQLite being tested */
@@ -4806,7 +4806,7 @@ func main1(tls *libc.TLS, argc int32, argv uintptr) (r int32) {
 		i++
 	}
 	if nHeap > 0 {
-		pHeap = libc.Xmalloc(tls, uint64(uint64(nHeap)))
+		pHeap = libc.Xmalloc(tls, uint64(nHeap))
 		if pHeap == uintptr(0) {
 			fatal_error(tls, __ccgo_ts+20093, libc.VaList(bp+16, nHeap))
 		}
@@ -4817,9 +4817,9 @@ func main1(tls *libc.TLS, argc int32, argv uintptr) (r int32) {
 	}
 	if doPCache != 0 {
 		if nPCache > 0 && szPCache > 0 {
-			pPCache = libc.Xmalloc(tls, uint64(int64(int64(nPCache))*int64(int64(szPCache))))
+			pPCache = libc.Xmalloc(tls, uint64(int64(nPCache)*int64(szPCache)))
 			if pPCache == uintptr(0) {
-				fatal_error(tls, __ccgo_ts+20154, libc.VaList(bp+16, int64(int64(nPCache))*int64(int64(szPCache))))
+				fatal_error(tls, __ccgo_ts+20154, libc.VaList(bp+16, int64(nPCache)*int64(szPCache)))
 			}
 		}
 		rc = libsqlite3.Xsqlite3_config(tls, int32(SQLITE_CONFIG_PAGECACHE), libc.VaList(bp+16, pPCache, szPCache, nPCache))
